src/zipfs_law/requirements.txt
pandas==2.2.2
matplotlib==3.9.0
python-dotenv==1.0.1
numpy==2.0.1